Fix: code organization, unified DB.
Still need to fix local DBMS.
